Workaround Netty bug by not use CompositeByteBuf
Due a bug in Netty [1] using a CompositeByteBuf can lead to data-corruption when partial writes happen on a socket. To workaround this bug we replaced the CompositeByteBuf usage (which was caused by Unpooled.wrappedBuffer(...)).

Beside work around the Netty bug this also reduce the object creation and so should stay in place even after Netty will be upgraded to 4.0.22.Final.

This commit also reverted the partial workaround that was in place before.

[1] https://github.com/netty/netty/issues/2761

patch by normanm; reviewed by tjake for CASSANDRA-7695