Fix an issue with transient dispatch failures
If a node goes down while hints are being dispatched to it,
the dispatch loop will not break until the node goes up again,
and dispatch is complete. This commit fixed the issue by propagating
failure upwards so that we can break the loop in case of failure.

patch by Aleksey Yeschenko; reviewed by Benedict Elliott Smith